Automatically backed up by Learn

+++ index.js
@@ -30,5 +30,5 @@ function makeInt(n)
 
 function preserveDecimal(n)
 {
-  return (parseInt(n))
+  return(parseFloat(n))
 }
